BB-2523 Add support for LTI embeds in course tabs and elsewhere
Extracting LTI Launch code from LTI XBlock functionality
